refactor: init rewrite config v2
