+++ .envrc
@@ -1,1 +1,2 @@
-export _NVIM_APPNAME=datsnvim
+export XDG_CONFIG_HOME=/home/dtsf/www/s/github.com/datsfilipe
+export NVIM_APPNAME=datsnvim
