+++ init.lua
@@ -1,8 +1,4 @@
-require 'core.unset'
-require 'core.options'
-require 'core.autocmds'
-require 'core.keymaps'
-require 'plugins.lazy'
+-- require 'dtsf.plugins.lazy'
 
-local config = require 'core.config'
-vim.cmd.colorscheme(config.colorscheme)
+local config = require 'utils.config'
+-- vim.cmd.colorscheme(config.colorscheme)
