--- lua/plugins/code/comment.lua
@@ -1,6 +1,0 @@
-return {
-  'echasnovski/mini.comment',
-  version = false,
-  event = 'BufRead',
-  opts = {},
-}
