--- lua/plugins/code/mini-pairs.lua
@@ -1,5 +1,0 @@
-return {
-  'echasnovski/mini.pairs',
-  version = false,
-  opts = {},
-}
