--- lua/plugins/code/refactoring.lua
@@ -1,17 +1,0 @@
-return {
-  'ThePrimeagen/refactoring.nvim',
-  event = 'BufRead',
-  keys = {
-    {
-      '<leader>r',
-      function()
-        require('refactoring').select_refactor()
-      end,
-      mode = 'v',
-      noremap = true,
-      silent = true,
-      expr = false,
-    },
-  },
-  opts = {},
-}
