--- lua/plugins/code/todo-comments.lua
@@ -1,6 +1,0 @@
-return {
-  'folke/todo-comments.nvim',
-  event = 'BufEnter',
-  dependencies = { 'nvim-lua/plenary.nvim' },
-  opts = {},
-}
