--- lua/plugins/code/tsc.lua
@@ -1,7 +1,0 @@
-return {
-  'dmmulroy/tsc.nvim',
-  opts = {
-    use_trouble_qflist = true,
-    run_as_monorepo = true,
-  },
-}
