--- lua/plugins/code/wakatime.lua
@@ -1,4 +1,0 @@
-return {
-  'wakatime/vim-wakatime',
-  event = 'VeryLazy',
-}
