--- lua/plugins/git/neogit.lua
@@ -1,33 +1,0 @@
-local keymap = vim.keymap
-keymap.set(
-  'n',
-  '<leader>gg',
-  ':Neogit<Return>',
-  { noremap = true, silent = true }
-)
-
-return {
-  'NeogitOrg/neogit',
-  cmd = 'Neogit',
-  dependencies = {
-    'nvim-lua/plenary.nvim',
-    'nvim-telescope/telescope.nvim',
-  },
-  config = function()
-    local neogit = require 'neogit'
-
-    neogit.setup {
-      disable_signs = false,
-      disable_context_highlighting = false,
-      disable_commit_confirmation = false,
-      signs = {
-        section = { '', '' },
-        item = { '', '' },
-        hunk = { '', '' },
-      },
-      integrations = {
-        diffview = true,
-      },
-    }
-  end,
-}
