--- lua/plugins/others/discord-presence.lua
@@ -1,4 +1,0 @@
-return {
-  'andweeb/presence.nvim',
-  event = 'BufEnter',
-}
