--- lua/plugins/others/gist.lua
@@ -1,4 +1,0 @@
-return {
-  'rawnly/gist.nvim',
-  cmd = { 'GistCreate', 'GistCreateFromFile', 'GistsList' },
-}
