--- lua/plugins/ui/telescope/extensions.lua
@@ -1,8 +1,0 @@
-return {
-  fzf = {
-    fuzzy = true,
-    override_generic_sorter = true,
-    override_file_sorter = true,
-    case_mode = 'smart_case',
-  },
-}
