+++ plugin/keymaps.lua
@@ -1,16 +1,10 @@
 vim.g.mapleader = ' '
 
-require('utils.discipline').cowboy()
-
 local keymap = vim.keymap
 local opts = { noremap = true, silent = true }
 
-keymap.set('n', 'ze', require('utils.excalidraw').open, opts)
-
--- ast-grep
-keymap.set('n', '<leader>A', require('utils.ast-grep').execute, opts)
-
-keymap.set('i', '<C-c>', '<Esc>', opts)
+keymap.set('n', '<leader>A', require('utils.ast-grep').execute, opts) -- ast-grep
+keymap.set('i', '<C-c>', '<Esc>', opts) -- everybody has a cursed mapping
 keymap.set('n', 'x', '"_x')
 
 -- avoid walking with space
@@ -30,40 +24,29 @@ keymap.set('v', 'K', ":m '<-2<Return>gv=gv", opts)
 keymap.set('n', '<leader>j', ':m .+1<Return>==', opts)
 keymap.set('n', '<leader>k', ':m .-2<Return>==', opts)
 
--- go back to last edited line
-keymap.set('n', 'g;', '`[', opts)
-
--- disable continuations
+keymap.set('n', 'g;', '`[', opts) -- go back to last edited line
 keymap.set('n', '<Leader>o', 'o<Esc>^Da', opts)
 keymap.set('n', '<Leader>O', 'O<Esc>^Da', opts)
 
--- new tab
 keymap.set('n', '<leader>t', ':tabnew<Return>', opts)
 keymap.set('n', '>', ':tabnext<Return>', opts)
 keymap.set('n', '<', ':tabprev<Return>', opts)
+
 -- split window
 keymap.set('n', '<leader>]', ':split<Return>', opts)
 keymap.set('n', '<leader>[', ':vsplit<Return>', opts)
--- maximize window
-keymap.set('n', '<leader>-', '<C-w>_<C-w><Bar>', opts)
--- equalize windows
-keymap.set('n', '<leader>=', '<C-w>=', opts)
+keymap.set('n', '<leader>-', '<C-w>_<C-w><Bar>', opts) -- maximize window
+keymap.set('n', '<leader>=', '<C-w>=', opts) -- equalize windows
 
--- wrap lines
-keymap.set('n', '<A-z>', ':set wrap!<Return>', opts)
 -- indent lines
 keymap.set('v', '<', '<gv', opts)
 keymap.set('v', '>', '>gv', opts)
+keymap.set('n', '<A-z>', ':set wrap!<Return>', opts) -- wrap lines
 
--- copy all
-keymap.set('n', '<leader>Y', 'ggVG"+y', opts)
--- copy to clipboard
-keymap.set({ 'n', 'v' }, '<leader>y', [["+y]], opts)
-
--- delete but don't yank
-keymap.set({ 'n', 'v' }, '<leader>d', [["_d]], opts)
--- paste but don't yank the deleted text
-keymap.set('x', '<leader>p', [["_dP]], opts)
+keymap.set('n', '<leader>Y', 'ggVG"+y', opts) -- copy all
+keymap.set({ 'n', 'v' }, '<leader>y', [["+y]], opts) -- copy to clipboard
+keymap.set({ 'n', 'v' }, '<leader>d', [["_d]], opts) -- delete but don't yank
+keymap.set('x', '<leader>p', [["_dP]], opts) -- paste but don't yank the deleted text
 
 -- replace occurrences
 keymap.set(
